CodeGenMonad

Description

Code generation monad. This module provides an abstraction for code generation computations with support for generation of distinct names, e.g. for labels.

Code generation computation

data CG i x a

Code generation computation. Parameterised on the type of instructions and additional auxiliary information. One use of the auxiliary information is for additional separate output sections by instantiating with suitable disjoint union type.

emit :: i -> CG i x ()

Emit instruction

emitAux :: x -> CG i x ()

Emit auxiliary information

divert :: CG i x a -> CG i x a

Divert output from sub-computation to separate section

newName :: CG i x Name

Generate a distinct name (within a run)

runCG :: CG i x a -> (a, [i], [x])

Run a code generation computation