Diagnostics

Description

Diagnostic messages and computations (monad). The module provides a datatype for representing diagnostic messages, tagged with source-code position and severity level (error, warning, information). It also provides an abstraction for diagnostic computations, where a computation has the ability to emit diagnostic messages and, possibly, fail. Finally, it provides a utility function for reporting internal compiler errors in a standardized way.

Diagnostic messages

data DMsg

Representation of diagnostic messages

Constructors

DMsg

Fields

dmLvl :: DMLvl

Severity level

dmSrcPos :: SrcPos

Associated source position

dmTxt :: String

Message text

data DMLvl

Severity levels

Constructors

DMLInfo
Information (least severe)

DMLWng
Warning

DMLErr
Error (most severe)

mkInfoMsg :: SrcPos -> String -> DMsg

Constructs information message.

mkWngMsg :: SrcPos -> String -> DMsg

Constructs warning message.

mkErrMsg :: SrcPos -> String -> DMsg

Constructs error message.

dmIsInfo :: DMsg -> Bool

Checks if information message.

dmIsWng :: DMsg -> Bool

Checks if warning message.

dmIsErr :: DMsg -> Bool

Checks if error message.

dmCmpLvlPos :: DMsg -> DMsg -> Ordering

Comparison function for ordering messages first by severity, then by source position.

ppDMsg :: DMsg -> String

Formats a diagnostic message for printing.

Diagnostic computations

class Monad d => Diagnostic d where

Class for diagnostic computations. Diagnostic computations accumulate diagnostic messages.

Methods

emitD :: DMsg -> d ()

Emits a diagnostic message

emitInfoD :: SrcPos -> String -> d ()

Emits an information message.

emitWngD :: SrcPos -> String -> d ()

Emits a warning message.

emitErrD :: SrcPos -> String -> d ()

Emits an error message.

getDMsgsD :: d [DMsg]

Diagnostic messages emitted thus far

(|||) :: d a -> d a -> d a

Tries the first diagnostic computation. If that results in errors, discards those and runs the second computation.

class Diagnostic df => DiagnosticFail df where

Class for diagnostic computations with failure.

Methods

failD :: SrcPos -> String -> df a

Emits an error message and fails.

failNoReasonD :: df a

Fails without giving any specific reason.

failIfErrorsD :: df ()

Fails if there has been errors thus far

stopD :: df a

Forces a stop, e.g. after some user-specified pass.

data D a

Diagnostic computation.

runD :: D a -> (a, [DMsg])

Runs a diagnostic computation. Returns:

Result of the computation.

Sorted list of diagnostic messages.

data DF a

dToDF :: D a -> DF a

runDF :: DF a -> (Maybe a, [DMsg])

Reporting and checking for internal errors

internalError :: String -> String -> String -> a

Signals an internal compiler error. Call with module name, name of function, and error message to report internal errors; i.e., things that should not happen.

assert :: Bool -> String -> String -> String -> a -> a

Checks that a condition holds and returns given value if so. Signals an internal compiler error otherwise. Call with condition, module name, name of function, error message, and value to return if condition satisfied.

assertM :: Monad m => Bool -> String -> String -> String -> m ()

Checks that a condition holds. Signals an internal compiler error otherwise. Call with condition, module name, name of function, error message.