Env

Description

Environment with operations. The module provides an abstract datatype for representing environments (symbol tables), along with operations for creating and extending environments and looking up symbols. There are three kinds of symbols: term-level symbols, which are internal or external, denoting entities like variables, constants, and procedures, and type-level symbols, denoting types.

Documentation

data Env

Environment (symbol table). Abstract.

mkTopLvlEnv :: [(Name, Type)] -> [(Name, Type, ExtSymVal)] -> Env

Creates an initial, top-level, environment of external symbols. Arguments:

List of name and type pairs for the type-level part of the top-level environment; i.e. a definition for each symbol.

List of name and type pairs for the term-level part of the top-levelenvironment; i.e., a declaration for each symbol.

Returns: A top-level environment.

openMajScope :: Env -> Env

Opens a new major (e.g. procedure/function) scope level.

openMinScope :: Env -> Env

Opens a new minor (e.g. let) scope level.

enterIntTermSym :: Name -> Type -> SrcPos -> Env -> Either IntTermSym (Env, IntTermSym)

Enters an internal term-level symbol into an environment. Enforces that symbols are uniquely defined at each scope level. Arguments:

Name of symbol to be entered.

Type of symbol to be entered.

Source position of the declaration or definition of symbol to be entered.

The environment to extend.

On success (Right), returns:

Extended environment.

Copy of the new symbol.

On failure (Left), returns:

The internal symbol with which there was a clash.

lookupTypeSym :: Name -> Env -> Maybe TypeSym

Looks up a type-level symbol. Arguments:

Name of symbol to lookup.

The environment in which to lookup the symbol.

On success, returns:

The symbol.

lookupTermSym :: Name -> Env -> Maybe TermSym

Looks up a term-level symbol. Later declarations (higher scope levels) shadow earlier ones. Arguments:

Name of symbol to lookup.

The environment in which to lookup the symbol.

On success, returns:

The symbol.

dummyTmS :: Name -> TermSym

Constructs a dummy (internal) term-level symbol to be used as a placeholder for example where a lookup has failed but a term-level symbol is needed.

Name of the symbol.

Returns:

The dummy symbol.