MTIR

Description

MiniTriangle Internal Representation. The definitions mirror the AST, but the program should be type correct at this stage and semantical annotations have been added. In particular, variables are represented by symbols (with name, type info, and source position) as opposed to just their names.

Documentation

data MTIR

Internal representation of MiniTriangle Programs

Constructors

MTIR

Fields

mtirCmd :: Command

data Command

Internal representation of syntactic category Command

Constructors

CmdAssign
Assignment

Fields

caVar :: Expression

Assigned variable

caVal :: Expression

Right-hand side expression

cmdSrcPos :: SrcPos

CmdCall
Procedure call

Fields

ccProc :: Expression

Called procedure

ccArgs :: [Expression]

Arguments

cmdSrcPos :: SrcPos

CmdSeq
Command sequence (block)

Fields

csCmds :: [Command]

Commands

cmdSrcPos :: SrcPos

CmdIf
Conditional command

Fields

ciCond :: Expression

Condition

ciThen :: Command

Then-branch

ciElse :: Command

Else-branch

cmdSrcPos :: SrcPos

CmdWhile
While-loop

Fields

cwCond :: Expression

Loop-condition

cwBody :: Command

Loop-body

cmdSrcPos :: SrcPos

CmdLet
Let-command

Fields

clDecls :: [Declaration]

Declarations

clBody :: Command

Let-body

cmdSrcPos :: SrcPos

data Expression

Internal representation of syntactic category Expression

Constructors

ExpLitBool
Literal Boolean

Fields

elbVal :: Bool

Literal Boolean.

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpLitInt
Literal integer

Fields

eliVal :: MTInt

Literal integer.

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpLitChr
Literal character

Fields

elcVal :: MTChr

Literal character.

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpExtRef
External reference (procedure/function)

Fields

eerVal :: Name

Name of external entity.

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpVar
Variable reference

Fields

evVar :: IntTermSym

Referenced variable (symbol!)

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpDeref
Dereferencing of reference variable

Fields

edArg :: Expression

Argument

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpApp
Function or n-ary operator application

Fields

eaFun :: Expression

Applied function or operator

eaArgs :: [Expression]

Arguments

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpAry
Array expression

Fields

eaElts :: [Expression]

Array elements

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

ExpIx
Array indexing

Fields

eiAry :: Expression

Array expression

eiIx :: Expression

Indexing expression

expType :: Type

Type (of application)

Type (after dereferencing)

Type

Type

Type

Type

Type

expSrcPos :: SrcPos

data Declaration

Internal representation of syntactic category Declaration (Most of these are also definitions. Note that symbols carry the types.)

Constructors

DeclConst
Constant declaration

Fields

dcConst :: IntTermSym

Declared constant (symbol!)

dcVal :: Expression

Value of defined constant

DeclVar
Variable declaration

Fields

dvVar :: IntTermSym

Declared variable (symbol!)

dvMbVal :: Maybe Expression

Initial value of declared variable, if any

DeclFun
Function declaration

Fields

dfFun :: IntTermSym

Declared function (symbol!)

dfArgs :: [IntTermSym]

Formal arguments (with types)

dfBody :: Expression

Function boody

DeclProc
Procedure declaration

Fields

dpProc :: IntTermSym

Declared procedure (symbol!)

dpArgs :: [IntTermSym]

Formal arguments (with types)

dpBody :: Command

Procedure boody

type MTInt = Int32

Haskell representaion of MiniTriangle integer values.

type MTChr = Char

Haskell representation of MiniTriangle character values