Symbol

Description

Symbols: information about named entities. Symbols are used to keep track of information about defined and declared named entities, like variables, procedures, functions, and types. They are used in the symbol table and MTIR.

There are two kinds of symbols: term-level symbols, denoting entities like variables and procedures, and type-level symbols, denoting types.

Documentation

data ExtSymVal

Constructors

ESVBool Bool

ESVInt MTInt

ESVChar Char

ESVLbl Name

type TermSym = Either ExtTermSym IntTermSym

Term-level symbol. Symbol denoting a term-level named entity, like a procedures, variable, or operator.

data ExtTermSym

Constructors

ExtTermSym
External symbol: defined outside current compilation unit

Fields

etmsName :: Name

Name

etmsType :: Type

Type of the symbol

etmsVal :: ExtSymVal

Value of the symbol

data IntTermSym

Constructors

IntTermSym
Internal symbol: defined in current compilation unit

Fields

itmsLvl :: ScopeLvl

Scope level

itmsName :: Name

Name

itmsType :: Type

Type of the symbol

itmsSrcPos :: SrcPos

Source code position of the declaration or definition.

data TypeSym

Type-level symbol. Symbols denoting a type. Can only be defined at the top level.

Constructors

TypeSym

Fields

tpsName :: Name

Name

tpsType :: Type

Representation of the type

tpsSrcPos :: SrcPos

Source code position of the type definition