Type

Description

MiniTriangle type representation and related definitions.

Documentation

data Type

Type for representing MiniTriangle types. Type representations can be compared for equality. Types are only equal if they have the same representation, except that SomeType is treated as equal to any type strictly as an implementation mechanism to avoid follow-on errors from a detected type error.

Constructors

SomeType
Some unknown type

Void
The empty type (return type of procedures)

Boolean
The Boolean type

Integer
The Integer type

Src Type
Read-only variable reference (source)

Snk Type
Write-only variable reference (sink)

Ref Type
Variable reference

Ary Type MTInt
Array; fields repr. element type and size

Arr [Type] Type
Type of procedures and functions (arrow). The fields represent the argument types and the result type. The latter is Void for procedures.

(<:) :: Type -> Type -> Bool

MiniTriangle Subtyping relation.

refType :: Type -> Bool

Predicate that decides if the type is a reference type.

rcdType :: Type -> Type

Referenced type: type of the reference

mapRcdType :: (Type -> Type) -> Type -> Type

Map referenced type (Is there a more principled approach?)

eltType :: Type -> Type

arySize :: Type -> MTInt

retType :: Type -> Type

Return type of a function type

type MTInt = Int32

Haskell representaion of MiniTriangle integer values.

isMTInt :: Integer -> Bool

type MTChr = Char

Haskell representation of MiniTriangle character values

isMTChr :: Char -> Bool